Add nav menu list and functional menu toggle
Add nav to About and Nasihat Bot. Both going to be modal only, not page.
Add js and css to make menu-toggle button show and hide nav.

Because About modal is where I can give a little talk and acknowledgements
to others that help and inspire me to do this. Nasihat Bot is modal to
users another benefit feature of this web.